Rename the quiz feature folder to quiz-build
