+++ lib/src/build_quiz/build_quiz_view_model.dart
@@ -1,8 +1,8 @@
 import 'package:stacked/stacked.dart';
 import 'package:kana_to_kanji/src/locator.dart';
-import 'package:kana_to_kanji/src/quiz/constants/alphabets.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/repositories/groups_repository.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/alphabets.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/repositories/groups_repository.dart';
 
 class BuildQuizViewModel extends FutureViewModel {
   final GroupsRepository _groupsRepository = locator<GroupsRepository>();
