+++ lib/src/build_quiz/constants/kana_type.dart
