+++ lib/src/build_quiz/models/kana.dart
@@ -1,7 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:kana_to_kanji/src/quiz/constants/alphabets.dart';
-import 'package:kana_to_kanji/src/quiz/constants/kana_type.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/alphabets.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
 
 part 'kana.freezed.dart';
