+++ lib/src/build_quiz/widgets/alphabet_groups_expansion_tile.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:kana_to_kanji/src/quiz/constants/alphabets.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/widgets/group_card.dart';
-import 'package:kana_to_kanji/src/quiz/widgets/kana_groups.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/alphabets.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/widgets/group_card.dart';
+import 'package:kana_to_kanji/src/build_quiz/widgets/kana_groups.dart';
 
 class AlphabetGroupsExpansionTile extends StatelessWidget {
   final Alphabets alphabet;
