+++ lib/src/build_quiz/widgets/group_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
 
 class GroupCard extends StatelessWidget {
   final Group group;
