+++ lib/src/build_quiz/widgets/kana_groups.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:kana_to_kanji/src/quiz/constants/kana_type.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/widgets/kana_groups_section.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/kana_type.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/widgets/kana_groups_section.dart';
 
 class KanaGroups extends StatelessWidget {
   final List<Group> groups;
