+++ lib/src/build_quiz/widgets/kana_groups_section.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:kana_to_kanji/src/quiz/constants/kana_type.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/widgets/group_card.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/kana_type.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/widgets/group_card.dart';
 
 class KanaGroupsSection extends StatelessWidget {
   final KanaTypes type;
