+++ lib/src/router.dart
@@ -1,5 +1,5 @@
 import 'package:go_router/go_router.dart';
-import 'package:kana_to_kanji/src/quiz/build_quiz_view.dart';
+import 'package:kana_to_kanji/src/build_quiz/build_quiz_view.dart';
 
 final router = GoRouter(initialLocation: BuildQuizView.routeName, routes: [
   GoRoute(
