+++ test/src/build_quiz/repositories/groups_repository_test.dart
@@ -1,8 +1,8 @@
 import 'package:kana_to_kanji/src/locator.dart';
-import 'package:kana_to_kanji/src/quiz/constants/alphabets.dart';
-import 'package:kana_to_kanji/src/quiz/constants/group_queries.dart';
-import 'package:kana_to_kanji/src/quiz/models/group.dart';
-import 'package:kana_to_kanji/src/quiz/repositories/groups_repository.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/alphabets.dart';
+import 'package:kana_to_kanji/src/build_quiz/constants/group_queries.dart';
+import 'package:kana_to_kanji/src/build_quiz/models/group.dart';
+import 'package:kana_to_kanji/src/build_quiz/repositories/groups_repository.dart';
 import 'package:mockito/annotations.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:kana_to_kanji/src/core/services/database_service.dart';
